requirements.txt
absl-py==2.1.0
accelerate==1.6.0
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
astor==0.8.1
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
audiomentations==0.40.0
audioread==3.0.1
beautifulsoup4==4.13.4
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
clearvoice==0.1.1
click==7.1.2
clone==0.0
cloudpickle==3.1.1
cn2an==0.5.23
colorama==0.4.6
contourpy==1.3.1
crcmod==1.7
cryptography==43.0.3
cycler==0.12.1
Cython==3.0.12
dashscope==1.23.1
dask==2025.2.0
datasets==2.8.0
decorator==5.2.1
dill==0.3.6
distro==1.9.0
editdistance==0.8.1
einops==0.8.1
exceptiongroup==1.2.2
executing==2.2.0
faiss-cpu==1.11.0
ffmpeg-python==0.2.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.6.0
fsspec==2024.12.0
funasr==1.2.6
future==1.0.0
gast==0.4.0
gdown==5.2.0
gensim==4.3.3
git-filter-repo==2.47.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-genai==1.10.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
h2==4.2.0
h5py==3.13.0
hdbscan==0.8.40
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.30.1
humanfriendly==10.0
hydra-core==1.3.2
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.6.1
ipython==8.34.0
jaconv==0.4.0
jamo==0.4.1
jax==0.4.38
jaxlib==0.4.38
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.6
jiter==0.9.0
jmespath==0.10.0
joblib==1.4.2
jsonplus==0.8.0
kaldi-native-fbank==1.21.1
kaldiio==2.18.1
keras==2.12.0
kiwisolver==1.4.8
lazy_loader==0.4
Levenshtein==0.27.1
libclang==18.1.1
librosa==0.10.2.post1
lightgbm==4.6.0
llvmlite==0.44.0
locket==1.0.0
loguru==0.7.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl-service==2.4.2
ml_dtypes==0.5.1
modelscope==1.25.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
multiprocess==0.70.14
namex==0.0.8
networkx==3.4.2
noisereduce==3.0.3
norbert==0.2.1
numba==0.61.0
numpy==1.26.4
numpy-minmax==0.3.1
numpy-rms==0.4.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
openai==1.78.1
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.14.1
oss2==2.19.1
packaging==24.2
paddlepaddle-gpu==2.6.0
pandas==1.5.3
parso==0.8.4
partd==1.4.2
peft==0.15.1
pexpect==4.9.0
pillow==11.2.0
platformdirs==4.3.7
pooch==1.8.2
proces==0.1.7
prompt_toolkit==3.0.50
propcache==0.3.1
proto-plus==1.26.1
protobuf==4.25.7
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyahocorasick==2.1.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycorrector==1.1.2
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.3
pydantic_core==2.33.1
pydub==0.25.1
Pygments==2.19.1
pynndescent==0.5.13
pyparsing==3.2.3
pypinyin==0.54.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-Levenshtein==0.27.1
python-stretch==0.3.1
python_speech_features==0.6
pytorch-wpe==0.0.1
pytz==2025.1
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
responses==0.18.0
rfc3986==1.5.0
rich==13.9.4
rotary-embedding-torch==0.8.3
rsa==4.9.1
safetensors==0.5.3
scenedetect==0.6.6
scikit-learn==1.6.1
scipy==1.13.0
sentencepiece==0.2.0
simpleaudio==1.0.4
simplejson==3.20.1
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.7
soxr==0.5.0.post1
spleeter==2.4.2
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.12.1
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==3.0.1
threadpoolctl==3.6.0
tokenizers==0.21.1
tomli==2.2.1
toolz==1.0.0
torch==2.1.0
torch-complex==0.4.4
torchaudio==2.1.0
torchinfo==1.8.0
torchvision==0.16.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.50.1
triton==2.1.0
typer==0.3.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.1
umap==0.1.1
umap-learn==0.5.7
uritemplate==4.1.1
urllib3==2.3.0
volcengine-python-sdk==1.1.5
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wordcloud==1.9.4
wrapt==1.14.1
xgboost==3.0.0
xxhash==3.5.0
yamlargparse==1.31.1
yapf==0.43.0
yarl==1.20.0
zipp==3.21.0
zstandard==0.23.0
